requirements.txt
minio
pyyaml

